stir/requirements.txt
absl-py==1.4.0
agate==1.6.3
aiohttp==3.8.4
aiohttp-cors==0.7.0
aiorwlock==1.3.0
aiosignal==1.3.1
airflow-clickhouse-plugin==1.0.0
airflow-dbt-python==0.15.2
albumentations==1.3.0
alembic==1.9.1
amqp==5.1.1
antlr4-python3-runtime==4.9.3
anyio==3.6.2
apache-airflow==2.6.3
apache-airflow-providers-celery==3.1.0
apache-airflow-providers-common-sql==1.3.1
apache-airflow-providers-ftp==3.2.0
apache-airflow-providers-http==4.1.0
apache-airflow-providers-imap==3.1.0
apache-airflow-providers-postgres==5.4.0
apache-airflow-providers-slack==7.3.2
apache-airflow-providers-sqlite==3.3.1
apache-airflow-providers-ssh==3.4.0
apispec==5.2.2
argcomplete==2.0.0
asgiref==3.7.2
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
Augmentor==0.2.10
Automat==20.2.0
Babel==2.11.0
backoff==2.2.1
bcrypt==4.0.1
better-exceptions==0.3.3
billiard==3.6.4.0
blessed==1.20.0
blinker==1.5
bottle==0.12.25
Brotli==1.0.9
cached-property==1.5.2
cachelib==0.9.0
cachetools==5.3.0
cattrs==22.2.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
clickclick==20.10.2
clickhouse-connect==0.5.12
clickhouse-driver==0.2.5
cloud-init==23.1.2
cmake==3.27.0
colorama==0.4.5
colorful==0.5.5
colorlog==4.8.0
command-not-found==0.3
commonmark==0.9.1
configobj==5.0.6
ConfigUpdater==3.1.1
connexion==2.14.1
constantly==15.1.0
contourpy==1.1.0
cron-descriptor==1.2.32
croniter==1.3.8
cryptography==39.0.0
cycler==0.11.0
datasets==2.13.1
dbt-clickhouse==1.3.2
dbt-core==1.3.1
dbt-extractor==0.4.1
dbt-postgres==1.3.1
dbus-python==1.2.18
decorator==4.4.2
Deprecated==1.2.13
dill==0.3.6
distlib==0.3.6
distro==1.7.0
distro-info===1.1build1
dnspython==2.2.1
docutils==0.19
email-validator==1.3.0
et-xmlfile==1.1.0
evaluate==0.4.0
eventlet==0.33.2
exceptiongroup==1.1.0
fastapi==0.95.2
filelock==3.9.0
Flask==2.2.2
Flask-AppBuilder==4.3.1
Flask-Babel==2.0.0
Flask-Caching==2.0.1
Flask-JWT-Extended==4.4.4
Flask-Limiter==3.5.0
Flask-Login==0.6.2
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
flatbuffers==23.3.3
flower==1.2.0
fonttools==4.40.0
frozenlist==1.3.3
fsspec==2023.5.0
future==0.18.2
gast==0.4.0
gensim==4.2.0
gevent==22.10.2
google-api-core==2.11.0
google-auth==2.17.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
google-re2==1.1
googleapis-common-protos==1.59.0
gpustat==1.1
graphviz==0.20.1
greenlet==2.0.1
grpcio==1.54.0
gunicorn==20.1.0
gyp==0.1
h11==0.14.0
h5py==3.8.0
hologram==0.0.15
httpcore==0.16.3
httplib2==0.20.2
httpx==0.23.2
huggingface-hub==0.15.1
humanize==4.4.0
hyperlink==21.0.0
idna==3.4
imagedominantcolor==1.0.1
ImageHash==4.3.1
imageio==2.31.1
imageio-ffmpeg==0.4.8
importlib-metadata==4.13.0
importlib-resources==6.0.1
incremental==21.3.0
inflection==0.5.1
isodate==0.6.1
itsdangerous==2.1.2
jax==0.4.13
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.2.0
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
keras==2.11.0
keyring==23.5.0
kiwisolver==1.4.4
kombu==5.2.4
langid==1.1.6
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-object-proxy==1.8.0
leather==0.3.4
libclang==16.0.0
limits==3.6.0
linkify-it-py==2.0.0
lit==16.0.6
load-dotenv==0.1.0
lockfile==0.12.2
Logbook==1.5.3
loguru==0.7.0
lz4==4.3.2
Mako==1.2.4
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.1
marshmallow==3.19.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
mashumaro==3.0.4
matplotlib==3.7.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
minimal-snowplow-tracker==0.0.2
ml-dtypes==0.2.0
more-itertools==8.10.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.4
multidict==6.0.4
multiprocess==0.70.14
mutagen==1.46.0
netifaces==0.11.0
networkx==2.8.8
nltk==3.8.1
numpy==1.21.6
numpyencoder==0.3.0
nvidia-ml-py==11.525.112
oauthlib==3.2.2
omegaconf==2.3.0
opencensus==0.11.2
opencensus-context==0.1.3
opencv-python==4.6.0.66
opencv-python-headless==4.7.0.72
openpyxl==3.1.2
opentelemetry-api==1.15.0
opentelemetry-exporter-otlp==1.15.0
opentelemetry-exporter-otlp-proto-grpc==1.15.0
opentelemetry-exporter-otlp-proto-http==1.15.0
opentelemetry-proto==1.15.0
opentelemetry-sdk==1.15.0
opentelemetry-semantic-conventions==0.36b0
opt-einsum==3.3.0
ordered-set==4.1.0
packaging==21.3
pandas==1.3.5
pandass==1.11.4
paramiko==3.0.0
parsedatetime==2.4
pathspec==0.9.0
pendulum==2.1.2
pexpect==4.8.0
pika==1.3.2
Pillow==9.5.0
platformdirs==2.6.2
pluggy==1.0.0
prison==0.2.1
proglog==0.1.10
prometheus-client==0.17.0
prompt-toolkit==3.0.36
protobuf==3.19.6
psutil==5.9.4
psycopg2==2.9.5
psycopg2-binary==2.9.5
ptyprocess==0.7.0
py-spy==0.3.14
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodomex==3.17
pydantic==1.10.8
Pygments==2.14.0
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.6.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==3.0.9
pyrsistent==0.19.3
pyserial==3.5
python-apt==2.3.0+ubuntu2.1
python-daemon==3.0.1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-dotenv==0.21.0
python-magic==0.4.24
python-nvd3==0.15.0
python-slugify==7.0.0
pytimeparse==1.1.8
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
pytzdata==2020.1
PyWavelets==1.3.0
PyYAML==6.0
qudida==0.0.4
ray==2.4.0
referencing==0.30.2
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rich==13.0.0
rich-argparse==1.3.0
rpds-py==0.10.3
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
SecretStorage==3.3.1
sentencepiece==0.1.99
service-identity==18.1.0
setproctitle==1.3.2
six==1.16.0
sklearn==0.0
slack-sdk==3.21.3
smart-open==6.3.0
sniffio==1.3.0
sos==4.4
SQLAlchemy==1.4.45
SQLAlchemy-JSONField==1.0.1.post0
SQLAlchemy-Utils==0.39.0
sqlparse==0.4.3
ssh-import-id==5.11
sshtunnel==0.4.0
starlette==0.27.0
swagger-ui-bundle==0.0.9
sympy==1.12
systemd-python==234
tabulate==0.9.0
tenacity==8.1.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-cpu-aws==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.1.1
text-unidecode==1.3
textblob==0.17.1
threadpoolctl==3.1.0
tifffile==2021.11.2
tokenizers==0.13.3
torch==2.0.1
tornado==6.1
tqdm==4.65.0
transformers==4.29.2
Twisted==22.1.0
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
ubuntu-advantage-tools==27.12
uc-micro-py==1.0.1
ufw==0.36.1
unattended-upgrades==0.1
unicodecsv==0.14.1
urllib3==1.26.13
uvicorn==0.22.0
videohash==3.0.1
vine==5.0.0
virtualenv==20.17.1
wadllib==1.3.6
wcwidth==0.2.6
websockets==11.0.1
Werkzeug==2.2.2
wrapt==1.14.1
WTForms==3.0.1
xlrd==1.2.0
xxhash==3.2.0
yarl==1.9.2
yt-dlp==2023.3.4
zipp==3.15.0
zope.event==4.5.0
zope.interface==5.5.2
zstandard==0.20.0
